Add support for account-level IP range ACLs
Any metadata attached to accounts with a key beginning with
'x-account-meta-iprange' will be treated as prefixes that should
be allowed access to the account and its containers.

Also restricted IP ACL-related metadata modification on accounts
and containers to only users with an 'admin' role. Additional
allowed roles may be specified using the 'meta_write_roles'
config option.

Updated README to cover account ACLs and metadata modification
limitations.

Bumped version to 1.0.1.